fixed directory setup

+++ Workspace/Freds-Spring-RPG-Engine/src/main/java/com/action/ActionRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActionRepo extends JpaRepository<Action, Integer> {
+public interface ActionRepo extends JpaRepository<Action, Long> {
 
 }
